+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## UNRELEASED
 
+### FEATURES
+
+* Yorc supports Slurm Accounting ([GH-280](https://github.com/ystia/yorc/issues/280))
+
 ## 3.2.0-M2 (February 15, 2019)
 
 ### DEPENDENCIES